Break fluent modifier cycle
Move the fluent parser implementation into modifiers.ts so modifier
functions can return FluentParser values without importing from fluent.ts.
The fluent subpath now re-exports the same API, and the fluent marker uses a
shared Symbol.for key for repeated decoration checks across loaded copies.

https://github.com/dahlia/optique/pull/843#discussion_r3465946152
https://github.com/dahlia/optique/pull/843#discussion_r3465946158
https://github.com/dahlia/optique/pull/843#discussion_r3465946164
